Adding railtie to enable twitter_cldr rake tasks in rails
