+++ lib/twitter_cldr/js.rb
@@ -11,15 +11,21 @@ require 'uglifier'
 require 'coffee-script'
 require 'rake'
 
-require 'twitter_cldr/js/tasks'
 require 'twitter_cldr/js/version'
 
 module TwitterCldr
   module Js
-    autoload :Compiler, "twitter_cldr/js/compiler"
+    autoload :Compiler,  "twitter_cldr/js/compiler"
     autoload :Renderers, "twitter_cldr/js/renderers"
+    autoload :Tasks,     "twitter_cldr/js/tasks/tasks"
 
     class Engine < ::Rails::Engine
+    end
+
+    class Railtie < ::Rails::Railtie
+      rake_tasks do
+        load "twitter_cldr/js/tasks/tasks.rake"
+      end
     end
   end
 end
